requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
requests==2.32.2
httpx==0.27.0
python-dotenv==1.0.1
streamlit==1.35.0
pydantic==2.7.1
